fee mastser, staff directory and alert setting finalized

+++ app/Http/Controllers/Resource/FeeGroupController.php
@@ -55,6 +55,10 @@ class FeeGroupController extends Controller
 
     public function destroy(FeeGroup $feeGroup)
     {
+        if ($feeGroup->feeTypes()->exists()) {
+            return back()->with(['failedMessage' => 'This Fee Group is being used and cannot be deleted.']);
+        }
+
         $feeGroup->delete();
 
         return back()->with('deleted', 'Deleted Successfully.');
