+++ app/Http/Controllers/Resource/StudentGroupController.php
@@ -55,6 +55,10 @@ class StudentGroupController extends Controller
 
     public function destroy(StudentGroup $studentGroup)
     {
+        if ($studentGroup->students()->exists()) {
+            return back()->with(['failedMessage' => 'This Student Group is being used and cannot be deleted.']);
+        }
+
         $studentGroup->delete();
 
         return back()->with('deleted', 'Deleted Successfully.');
